pyproject.toml
[tool.poetry]
name = "labs"
version = "0.1.0"
description = ""
authors = ["Maksym Yatskov <yatskov.maksym@lll.kpi.ua>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
